Add initial implementation of LadderGame and Ladder
Ladder's internal state is the same as how it looks, which is not good but
is enough to meet step-1's requirement.

The generation process thus also works on raw string.

The generation method is named so, because the method is naive and we will
probably need another generation method very soon.