[frank] 1단계 리뷰 반영 구현 (#272)
* Remove and untrack the .idea folder

This is a suggestion from #137.

* Improve generation performance

Generation method now use a static Random object instead of Instantiating a
new Random object each time it is called.

* Remove unnecessary comment

Those comment is now implemented as code and the code is quite straightforward.

* Rename the ladder state variable with a more meaningful name

* Rewrite toString method

The method given the name toString has a common practice to respect, and it was
not in the old code.

The method to return the ladder's shape as string is given a new name, and the
toString method is rewritten to do what it is supposed to do.

* Add various code improvement

1. Ladder state is declared as List instead of ArrayList for better flexibility
2. Unnecessary "this." is removed
3. width, height is made final

Co-authored-by: frank.burger <frank.burger@kakaocorp.com>